Refactor command_line utils
